fixing xmlquery issues
Fixes to xmlquery and extensive new documentation

This PR provides fixes #946
Particularly the arguments --listall --full now work and --groups has been replaced by --get-groups and is now functional.

Test suite: scripts_regression_tests.py
Test baseline: NA
Test namelist changes: NA
Test status: NA

Fixes #946
User interface changes?: xmlquery --group has been changed go xmlquery --get-group
Code review: jedwards

+++ scripts/lib/CIME/case.py
@@ -321,7 +321,8 @@ class Case(object):
                     elif field == "group":
                         result.extend(env_file.get_groups(root))
                     elif field == "valid_values":
-                        vv = env_file.get_valid_values(variable)
+                        # pylint: disable=private-member
+                        vv = env_file._get_valid_values(root) 
                         if vv:
                             result.extend(vv)
                     elif field == "file":
@@ -1333,7 +1334,8 @@ class Case(object):
         """
         expect(os.path.isfile(xmlfile), "Could not find file {}".format(xmlfile))
 
-        self.flush(flushall=True)
+        if not self._read_only_mode:
+            self.flush(flushall=True)
 
         gfile = GenericXML(infile=xmlfile)
         ftype = gfile.get_id()
